requirements.txt
Flask==2.2.2
Flask-SQLAlchemy==3.0.2
Werkzeug==2.2.2
python-dotenv==1.0.0
